+++ src/lib/components/Lightbox.svelte
